+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v2.4.0 - Unrelased
+
+Added
+  * AZ-234: Add option to force Static private IP
+  * AZ-234: Add option to associate Network Security Group to the NIC
+  * AZ-234: Output NIC ID
+  * AZ-222: Option to activate `enable_accelerated_networking` on NIC resource
+
+Fixed
+  * AZ-167: Fix NIC configuration name
+
 # v2.3.1 - 2020-07-09
 
 Changed